Moving logs.go to server package because of cycle issues
